Correcciones de PR "Primer parte del indice - Contactanos"
Se hicieron las correcciones solicitadas por Blackguz en la primera versión del PR
